added opting to byzantine generals example; one now decides to do one of: - "attack" :: |packets| == |links| && attack packets > retreat packets - "retreat" :: |packets| == |links| && attack packets < retreat packets - "unknown" :: |packets| == |links| && attack packets == retreat packets - "not yet known" :: |packets| < |links|
this can be seen in the variables panel

+++ src/vidis/modules/byzantineGenerals/RetreatPacket.java
@@ -8,7 +8,7 @@ public class RetreatPacket extends APacket {
 	public RetreatPacket() {
 		super();
 	}
-	public RetreatPacket(double id) {
+	public RetreatPacket(int id) {
 		super(id);
 	}
